Fixed : IIBTree needs integer keys http://plone.org/products/alm.solrindex/issues/3

+++ alm/solrindex/index.py
@@ -278,7 +278,7 @@ class SolrIndex(PropertyManager, SimpleItem):
         uniqueKey = cm.schema.uniqueKey
         result = IIBTree()
         for r in response:
-            result[r[uniqueKey]] = int(r.get('score', 0) * 1000)
+            result[int(r[uniqueKey])] = int(r.get('score', 0) * 1000)
 
         return result, queried
